Extract Chart component from SearchFunnelWidgetGA4.
